move `is_jupyter` function to `utils`
