+++ drilldown/image/image.py
@@ -8,20 +8,7 @@ from aiohttp import web
 import large_image
 import uuid
 
-
-def is_jupyter():
-    from IPython import get_ipython
-
-    try:
-        shell = get_ipython().__class__.__name__
-        if shell == "ZMQInteractiveShell":
-            return True  # Jupyter notebook or qtconsole
-        elif shell == "TerminalInteractiveShell":
-            return False  # Terminal running IPython
-        else:
-            return False  # Other type (?)
-    except NameError:
-        return False
+from ..utils import is_jupyter
 
 
 class ImageViewer:
