+++ drilldown/plot/plotly_plots.py
@@ -8,22 +8,7 @@ from trame.ui.vuetify import SinglePageLayout
 from plotly import express as px
 import numpy as np
 
-from ..utils import convert_to_numpy_array
-
-
-def is_jupyter():
-    from IPython import get_ipython
-
-    try:
-        shell = get_ipython().__class__.__name__
-        if shell == "ZMQInteractiveShell":
-            return True  # Jupyter notebook or qtconsole
-        elif shell == "TerminalInteractiveShell":
-            return False  # Terminal running IPython
-        else:
-            return False  # Other type (?)
-    except NameError:
-        return False
+from ..utils import convert_to_numpy_array, is_jupyter
 
 
 class PlotlyPlot:
