+++ drilldown/ui/ui.py
@@ -11,21 +11,7 @@ import panel as pn
 from functools import partial
 
 from ..plotter import DrillDownPlotter
-
-
-def is_jupyter():
-    from IPython import get_ipython
-
-    try:
-        shell = get_ipython().__class__.__name__
-        if shell == "ZMQInteractiveShell":
-            return True  # Jupyter notebook or qtconsole
-        elif shell == "TerminalInteractiveShell":
-            return False  # Terminal running IPython
-        else:
-            return False  # Other type (?)
-    except NameError:
-        return False
+from ..utils import is_jupyter
 
 
 def ui_card(title, ui_name):
