+++ drilldown/utils.py
@@ -23,3 +23,18 @@ def convert_to_numpy_array(arr, collapse_dim=True):
         raise TypeError("Data must be a sequence or Pandas object.")
 
     return arr
+
+
+def is_jupyter():
+    from IPython import get_ipython
+
+    try:
+        shell = get_ipython().__class__.__name__
+        if shell == "ZMQInteractiveShell":
+            return True  # Jupyter notebook or qtconsole
+        elif shell == "TerminalInteractiveShell":
+            return False  # Terminal running IPython
+        else:
+            return False  # Other type (?)
+    except NameError:
+        return False
